Add misc fixes
