+++ src/tools/parsers.py
@@ -2,15 +2,6 @@ from src import db
 from src.repositories.survey_repository import survey_repository
 from sqlalchemy import text
 
-def clear_database():
-    db.session.execute(text("DELETE FROM final_group"))
-    db.session.execute(text("DELETE FROM user_survey_rankings"))
-    db.session.execute(text("DELETE FROM choice_infos"))
-    db.session.execute(text("DELETE FROM survey_choices"))
-    db.session.execute(text("DELETE FROM surveys"))
-    db.session.execute(text("DELETE FROM users"))
-    db.session.commit()
-
 def parser_elomake_csv(file, survey_name, user_id):
     '''
     Parses a survey from Elomake exported CSV file and creates a survey,
