+++ tests/survey_repository_test.py
@@ -6,7 +6,7 @@ from src import db
 from src.repositories.survey_repository import survey_repository as sr
 from src.repositories.user_repository import user_repository as ur
 from src.entities.user import User
-from src.tools.parsers import clear_database
+from src.tools.db_tools import clear_database
 
 class TestSurveyRepository(unittest.TestCase):
     def setUp(self):
